fix low-severity issues: typos, dead code, correctness
fusb302b:
- fusb302_defines: fix self-referential macro TOG_SAVE_PWR to use _SHIFT
- pd.cpp: remove dead code (always-true/false ternaries, unused power var)
- pd.h: remove duplicate break statement

satellite1:
- satellite1.cpp: remove unused arr debug variables

tas2780:
- fix typo "Dectivating" -> "Deactivating"
- replace static local last_call with per-instance member last_loop_check_

memory_flasher:
- fix typo bock_size -> block_size in virtual method parameter
- remove declared-but-never-defined cleanup_() method
- rename misleading wifi_connected_to_code -> in_progress_to_code

micro_wake_word:
- add missing f-string prefix in ValueError message
- rename partially_detection -> partial_detection (typo)

audio:
- chunked_ring_buffer.h: change uint8_t data[0] (GCC extension) to
  standard flexible array member data[]
- audio.h: fix signed integer overflow UB in tv_t::normalize() when
  usec == INT32_MIN by casting to int64_t before negation

config:
- led_ring.yaml: rename error_color -> success_color in Success effect
- satellite1.ld2410.yaml: fix misleading package key mmwave_ld2450 ->
  mmwave_ld2410
- satellite1.yaml: fix typo "orignal" -> "original" in comment

+++ esphome/components/audio/audio.h
@@ -213,7 +213,7 @@ struct tv_t {
       sec += usec / 1000000;
       usec %= 1000000;
     } else if (usec < 0) {
-      int32_t borrow = (-usec + 999999) / 1000000;
+      int32_t borrow = (static_cast<int64_t>(-usec) + 999999) / 1000000;
       sec -= borrow;
       usec += borrow * 1000000;
     }
